Fix comparer
